pyproject.toml
[project]
name = "llm-evaluation-with-mlflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dagshub>=0.5.10",
    "mlflow>=3.1.0",
    "numpy>=2.2.6",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
]
